Fix warning treated as error: warning C4244: 'initializing' : conversion from '__int64' to 'const long double', possible loss of data

+++ source/examples/common/ExampleWindowEventHandler.cpp
@@ -81,7 +81,7 @@ void ExampleWindowEventHandler::computeFps(glowwindow::PaintEvent & event)
 
            event.window()->setTitle(stream.str());
 
-           m_swapElapsedTime = m_timer.elapsed().count();
+           m_swapElapsedTime = static_cast<long double>(m_timer.elapsed().count());
            m_swapCount = 0;
        }
 }
